refactor: created custom hook for changing playback
