+++ src/components/Shuffle.tsx
@@ -1,6 +1,8 @@
-import { getAlbum, putPlayback } from '@/lib/spotifyApi';
-import { useMutation } from '@tanstack/react-query';
 import { useSession } from 'next-auth/react';
+import { redirect } from 'next/navigation';
+
+import usePlayback from '@/hooks/usePlayback';
+import { getAlbum } from '@/lib/spotifyApi';
 
 interface Props {
   resource: 'album';
@@ -9,21 +11,11 @@ interface Props {
 
 const Shuffle = ({ resource, amountOfResource }: Props) => {
   const { data: session } = useSession();
-  const mutation = useMutation({
-    mutationFn: ({
-      access_token,
-      uri,
-    }: {
-      access_token: string;
-      uri: string;
-    }) => {
-      return putPlayback(access_token, uri);
-    },
-  });
-
   if (!session) {
-    return;
+    redirect('/signin');
   }
+
+  const mutation = usePlayback();
 
   const handleClick = async () => {
     let uri = '';
